requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
chromadb>=0.4.0
pypdf2>=3.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.0.0,<3.0.0
streamlit>=1.28.0
tiktoken>=0.5.0
sentence-transformers>=2.2.0
openai>=1.0.0
